+++ packages/vue/src/components/PieChart.ts
@@ -24,11 +24,13 @@ import {
   PIE_BASE_SHADOW,
   PIE_EMPHASIS_SHADOW,
   getChartElementOpacity,
+  DEFAULT_DONUT_INNER_RADIUS_RATIO,
   type ChartLegendItem,
   type ChartLegendPosition,
   type ChartPadding,
   type PieChartDatum,
-  type PieChartProps as CorePieChartProps
+  type PieChartProps as CorePieChartProps,
+  type DonutChartProps as CoreDonutChartProps
 } from '@expcat/tigercat-core'
 import { ChartCanvas } from './ChartCanvas'
 import { ChartLegend } from './ChartLegend'
@@ -468,3 +470,46 @@ export const PieChart = defineComponent({
 })
 
 export default PieChart
+
+export interface VueDonutChartProps extends CoreDonutChartProps {
+  data: PieChartDatum[]
+  padding?: ChartPadding
+}
+
+export type DonutChartProps = VueDonutChartProps
+
+/**
+ * PieChart alias with a default inner radius. Prefer PieChart + `innerRadiusRatio` in new code.
+ */
+export const DonutChart = defineComponent({
+  name: 'TigerDonutChart',
+  inheritAttrs: false,
+  props: {
+    data: {
+      type: Array as PropType<PieChartDatum[]>,
+      required: true
+    },
+    innerRadiusRatio: {
+      type: Number,
+      default: DEFAULT_DONUT_INNER_RADIUS_RATIO
+    },
+    className: {
+      type: String
+    }
+  },
+  emits: ['update:hoveredIndex', 'update:selectedIndex', 'slice-click', 'slice-hover'],
+  setup(props, { attrs, emit }) {
+    return () =>
+      h(PieChart, {
+        ...attrs,
+        data: props.data,
+        innerRadiusRatio: props.innerRadiusRatio,
+        className: classNames(coerceClassValue(attrs.class), props.className),
+        'onUpdate:hoveredIndex': (index: number | null) => emit('update:hoveredIndex', index),
+        'onUpdate:selectedIndex': (index: number | null) => emit('update:selectedIndex', index),
+        onSliceClick: (index: number, datum: PieChartDatum) => emit('slice-click', index, datum),
+        onSliceHover: (index: number | null, datum: PieChartDatum | null) =>
+          emit('slice-hover', index, datum)
+      })
+  }
+})